git-svn-id: file:///var/svn-repos/e-head/trunk@426 2e978b77-0c5f-4b2b-b7d0-db7ae35eb1e0

+++ library/PMS/Staff.php
@@ -71,8 +71,35 @@ class PMS_Staff
             return $response;
         }
 
+        $file = $_FILES['cv_file'];
+
+        if ($file['error'] != UPLOAD_ERR_NO_FILE) {
+
+            if ($file['error'] > 0) {
+                $response->addStatus(new PMS_Status(
+                    PMS_Status::INPUT_PARAMS_INCORRECT, 'file'));
+                return $response;
+            }
+
+            $filenameArray = split('\.', $file['name']);
+            $ext = array_pop($filenameArray);
+            $filename = uniqid() . '.' . $ext;
+            $filepath = FILES_DIR . DIRECTORY_SEPARATOR . $filename;
+
+            if (move_uploaded_file($file['tmp_name'], $filepath)) {
+                $data = $f->getData();
+                $data['cv_file'] = $filename;
+                $f->setData($data);
+            } else {
+                return $response->addStatus(new PMS_Status(PMS_Status::FAILURE));
+            }
+        }
+
         $id = $this->_table->insert($f->getData());
         if (!$id) {
+            if (isset($filepath) && is_file($filepath)) {
+                unlink($filepath);
+            }
             return $response->addStatus(new PMS_Status(PMS_Status::DATABASE_ERROR));
         }
 
@@ -100,6 +127,41 @@ class PMS_Staff
         if ($response->hasNotSuccess()) {
             return $response;
         }
+
+        $file = $_FILES['cv_file'];
+
+        if ($file['error'] != UPLOAD_ERR_NO_FILE) {
+
+            if ($file['error'] > 0) {
+                $response->addStatus(new PMS_Status(
+                    PMS_Status::INPUT_PARAMS_INCORRECT, 'file'));
+                return $response;
+            }
+
+            $resp = $this->get($f->id);
+            if ($resp->hasNotSuccess()) {
+                return $response->importStatuses($resp->getStatusCollection());
+            }
+            $row = $resp->getRow();
+
+            $filenameArray = split('\.', $file['name']);
+            $ext = array_pop($filenameArray);
+            $filename = uniqid() . '.' . $ext;
+            $filepath = FILES_DIR . DIRECTORY_SEPARATOR . $filename;
+
+            if (move_uploaded_file($file['tmp_name'], $filepath)) {
+                $data = $f->getData();
+                $data['cv_file'] = $filename;
+                $f->setData($data);
+                if (!empty($row['cv_file'])
+                && is_file(FILES_DIR . DIRECTORY_SEPARATOR . $row['cv_file'])) {
+                    unlink(FILES_DIR . DIRECTORY_SEPARATOR . $row['cv_file']);
+                }
+            } else {
+                return $response->addStatus(new PMS_Status(PMS_Status::FAILURE));
+            }
+        }
+
         $rows = $this->_table->updateByPk($f->getData(), $f->id);
         $status = PMS_Status::retrieveAffectedRowStatus($rows);
         return $response->addStatus(new PMS_Status($status));
@@ -113,8 +175,22 @@ class PMS_Staff
             return $response->addStatus(new PMS_Status(
                 PMS_Status::INPUT_PARAMS_INCORRECT, 'id'));
         }
+
+        $resp = $this->get($id);
+        if ($resp->hasNotSuccess()) {
+            return $response->importStatuses($resp->getStatusCollection());
+        }
+        $row = $resp->getRow();
+
         $res = $this->_table->deleteByPk($id);
-        $status = false === $res ? PMS_Status::DATABASE_ERROR : PMS_Status::OK;
-        return $response->addStatus(new PMS_Status($status));
+        if (false === $res) {
+            return $response->addStatus(new PMS_Status(PMS_Status::DATABASE_ERROR));
+        }
+
+        if (!empty($row['cv_file'])
+        && is_file(FILES_DIR . DIRECTORY_SEPARATOR . $row['cv_file'])) {
+            unlink(FILES_DIR . DIRECTORY_SEPARATOR . $row['cv_file']);
+        }
+        return $response->addStatus(new PMS_Status(PMS_Status::OK));
     }
 }
